+++ Genesis/src/utopia/genesis/event/MouseWheelEvent.scala
@@ -9,9 +9,13 @@ import utopia.genesis.shape.shape2D.Point
  * @param wheelTurn The amount of mouse wheel 'notches' since the last event. A positive number
  * indicates a wheel turn towards the user. A negative number indicates a roll away from the user.
   * @param mousePosition Current mouse position.
+  * @param absoluteMousePosition Current mouse position in the screen coordinate system (in pixels)
   * @param buttonStatus Current mouse button status
+  * @param consumeEvent An event concerning this event's consuming. None if not consumed yet (default)
  */
-case class MouseWheelEvent(wheelTurn: Double, mousePosition: Point, buttonStatus: MouseButtonStatus,
+case class MouseWheelEvent(wheelTurn: Double, override val mousePosition: Point,
+						   override val absoluteMousePosition: Point,
+						   override val buttonStatus: MouseButtonStatus,
 						   override val consumeEvent: Option[ConsumeEvent] = None) extends MouseEvent[MouseWheelEvent]
 	with Consumable[MouseWheelEvent]
 {