+++ Reflection/src/utopia/reflection/container/swing/window/Popup.scala
@@ -2,14 +2,20 @@ package utopia.reflection.container.swing.window
 
 import java.awt.event.{WindowEvent, WindowFocusListener}
 
+import utopia.genesis.event.{KeyStateEvent, MouseButtonStateEvent}
+import utopia.genesis.handling.{KeyStateListener, MouseButtonStateListener}
 import utopia.reflection.localization.LocalString._
 import utopia.genesis.handling.mutable.ActorHandler
 import utopia.genesis.shape.shape2D.{Point, Size}
 import utopia.genesis.util.Screen
+import utopia.genesis.view.{GlobalKeyboardEventHandler, GlobalMouseEventHandler}
+import utopia.inception.handling.Mortal
+import utopia.inception.handling.immutable.Handleable
 import utopia.reflection.component.swing.template.AwtComponentRelated
 import utopia.reflection.component.template.ComponentLike
 import utopia.reflection.component.template.layout.stack.Stackable
 import utopia.reflection.container.swing.AwtContainerRelated
+import utopia.reflection.container.swing.window.Popup.PopupAutoCloseLogic.{Never, WhenAnyKeyPressed, WhenClickedOutside, WhenEscPressed, WhenFocusLost}
 import utopia.reflection.container.swing.window.WindowResizePolicy.Program
 import utopia.reflection.shape.Alignment
 import utopia.reflection.shape.Alignment.TopLeft
@@ -27,9 +33,10 @@ object Popup
 	  * @param content Popup contents (stackabe container)
 	  * @param actorHandler An actorhandler that will supply the pop-up with action events. These are required in
 	  *                     mouse- and keyboard event generating.
-	  * @param hideWhenFocusLost Whether the pop-up should be hidden when it loses focus (default = true)
 	  * @param resizeAlignment Alignment used for handling pop-up position when its size changes.
 	  *                        Default = Top Left = Pop up's position wil remain the same
+	  * @param autoCloseLogic Logic that determines when this pop-up is automatically closed
+	  *                       (default = never close automatically)
 	  * @param getTopLeft A function for calculating the new top left corner of the pop-up within context component.
 	  *                   Provided parameters are: a) Context size and b) pop-up window size
 	  * @tparam C Type of displayed item
@@ -37,15 +44,18 @@ object Popup
 	  */
 	def apply[C <: AwtContainerRelated with Stackable](context: ComponentLike with AwtComponentRelated, content: C,
 													   actorHandler: ActorHandler,
-													   hideWhenFocusLost: Boolean = true,
+													   autoCloseLogic: PopupAutoCloseLogic = Never,
 													   resizeAlignment: Alignment = TopLeft)
 													  (getTopLeft: (Size, Size) => Point) =
 	{
 		// If context isn't in a window (which it should), has to use a Frame instead of a dialog
 		val owner = context.parentWindow
 		val windowTitle = "Popup".local("en").localizationSkipped
-		val newWindow = owner.map { o => new Dialog(o, content, windowTitle, Program, resizeAlignment, borderless = true) }.getOrElse(
-			Frame.windowed(content, windowTitle, Program, borderless = true))
+		val newWindow = owner match
+		{
+			case Some(window) => new Dialog(window, content, windowTitle, Program, resizeAlignment, borderless = true)
+			case None => Frame.windowed(content, windowTitle, Program, borderless = true)
+		}
 		
 		// Calculates the absolute target position
 		val newPosition = context.absolutePosition + getTopLeft(context.size, newWindow.size)
@@ -54,17 +64,55 @@ object Popup
 		val maxPosition = (Screen.size - newWindow.size).toPoint
 		newWindow.position = newPosition topLeft maxPosition
 		
-		// Activates focus listening, if necessary
-		if (hideWhenFocusLost)
-			newWindow.component.addWindowFocusListener(new HideOnFocusLostListener(newWindow))
+		// Determines auto close logic
+		autoCloseLogic match
+		{
+			case WhenFocusLost => newWindow.component.addWindowFocusListener(new HideOnFocusLostListener(newWindow))
+			case WhenClickedOutside => GlobalMouseEventHandler += new HideOnOutsideClickListener(newWindow)
+			case WhenAnyKeyPressed => GlobalKeyboardEventHandler += new HideOnKeyPressListener(newWindow)
+			case WhenEscPressed => newWindow.setToCloseOnEsc()
+			case _ => ()
+		}
 		
 		newWindow.startEventGenerators(actorHandler)
-		newWindow.setToCloseOnEsc()
 		newWindow
 	}
 	
 	
 	// NESTED	----------------------
+	
+	/**
+	  * A common trait for different automated popup close logic options
+	  */
+	sealed trait PopupAutoCloseLogic
+	
+	object PopupAutoCloseLogic
+	{
+		/**
+		  * Popup is never closed automatically
+		  */
+		case object Never extends PopupAutoCloseLogic
+		
+		/**
+		  * Popup is closed automatically when it first loses focus
+		  */
+		case object WhenFocusLost extends PopupAutoCloseLogic
+		
+		/**
+		  * Popup is closed automatically when user clicks outside the popup area
+		  */
+		case object WhenClickedOutside extends PopupAutoCloseLogic
+		
+		/**
+		  * Popup is closed automatically when any key is pressed in any application window
+		  */
+		case object WhenAnyKeyPressed extends PopupAutoCloseLogic
+		
+		/**
+		  * Popup is closed automatically when esc is pressed while this popup has focus
+		  */
+		case object WhenEscPressed extends PopupAutoCloseLogic
+	}
 	
 	private class HideOnFocusLostListener(popup: Window[_]) extends WindowFocusListener
 	{
@@ -75,5 +123,35 @@ object Popup
 			popup.component.removeWindowFocusListener(this)
 			popup.close()
 		}
+	}
+	
+	private trait HideActionListener extends Handleable with Mortal
+	{
+		// ABSTRACT	--------------------------
+		
+		protected def popup: Window[_]
+		
+		
+		// IMPLEMENTED	----------------------
+		
+		override def isDead = popup.isClosed
+	}
+	
+	private class HideOnOutsideClickListener(override val popup: Window[_]) extends MouseButtonStateListener
+		with HideActionListener
+	{
+		override def onMouseButtonState(event: MouseButtonStateEvent) =
+		{
+			if (popup.isVisible && !popup.bounds.contains(event.absoluteMousePosition))
+				popup.close()
+			None
+		}
+	}
+	
+	private class HideOnKeyPressListener(override val popup: Window[_]) extends KeyStateListener with HideActionListener
+	{
+		override val keyStateEventFilter = KeyStateEvent.wasPressedFilter
+		
+		override def onKeyState(event: KeyStateEvent) = popup.close()
 	}
 }