+++ Inception/Changes.md
@@ -1,4 +1,10 @@
 # Utopia Inception - List of Changes
+## v2.3 (beta)
+### Breaking Changes
+- The Killable trait no longer contains any attributes in itself, only the definition of method kill()
+### New Methods
+- Added .debugString to Handler and HandlerRelay
+
 ## v2.2
 ### Scala
 - Module is now based on Scala v2.13.3